+++ tests/test_elevation_motion.py
@@ -23,6 +23,8 @@ import logging
 import math
 import unittest
 
+import pytest
+
 from lsst.ts import mtdome
 
 logging.basicConfig(
@@ -95,7 +97,7 @@ class ElevationMotionTestCase(unittest.IsolatedAsyncioTestCase):
             crawl_velocity=velocity,
             motion_state=motion_state,
         )
-        self.assertEqual(expected_duration, duration)
+        assert expected_duration == duration
 
     async def verify_elevation_motion(
         self,
@@ -123,9 +125,9 @@ class ElevationMotionTestCase(unittest.IsolatedAsyncioTestCase):
             velocity,
             motion_state,
         ) = self.elevation_motion.get_position_velocity_and_motion_state(tai)
-        self.assertAlmostEqual(expected_position, position)
-        self.assertAlmostEqual(expected_velocity, velocity)
-        self.assertEqual(expected_motion_state, motion_state)
+        assert expected_position == pytest.approx(position)
+        assert expected_velocity == pytest.approx(velocity)
+        assert expected_motion_state == motion_state
 
     async def test_move_zero_ten(self) -> None:
         """Test the ElevationMotion when moving from position 0 to